+++ packages/hamburger-react/src/index.ts
@@ -1,1 +1,2 @@
+export type { HamburgerProps } from "./Hamburger";
 export { Hamburger } from "./Hamburger";