+++ packages/message-box-react/src/index.ts
@@ -1,2 +1,4 @@
 export type { MessageBoxProps } from "./MessageBox";
 export { InfoMessageBox, ErrorMessageBox, SuccessMessageBox, WarningMessageBox } from "./MessageBox";
+export type { FormErrorMessageBoxProps } from "./FormErrorMessageBox";
+export { FormErrorMessageBox } from "./FormErrorMessageBox";