+++ packages/summary-table-react/src/index.ts
@@ -1,1 +1,2 @@
 export { SummaryTable } from "./SummaryTable";
+export { SummaryTableRow } from "./SummaryTableRow";