+++ packages/validators-util/src/index.ts
@@ -3,3 +3,8 @@ export { isValidTelefonnummer } from "./isValidTelefonnummer/isValidTelefonnumme
 export { isValidRegistreringsnummer } from "./isValidRegistreringsnummer/isValidRegistreringsnummer";
 export { isValidChassisnummer } from "./isValidChassisnummer/isValidChassisnummer";
 export { isValidDogId } from "./isValidDogId/isValidDogId";
+export { isExactLength } from "./isExactLength/isExactLength";
+export { isInteger } from "./isInteger/isInteger";
+export { isValidFodselsnummer } from "./isValidFodselsnummer/isValidFodselsnummer";
+export { isValidName } from "./isValidName/isValidName";
+export { hasMinimumWords } from "./hasMinimumWords/hasMinimumWords";