+++ portal/CHANGELOG.md
@@ -3,6 +3,27 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 11.0.0 (2021-12-22)
+
+### Bug Fixes
+
+- rett a11y-problemer i Hamburger ([721db9f](https://github.com/fremtind/jokul/commit/721db9f6632f523e2b38badbc0b54eddecf868bb)), closes [#2606](https://github.com/fremtind/jokul/issues/2606)
+
+### BREAKING CHANGES
+
+- - Fjernet `description`-propen. Den var en renamet `aria-label`. Bruk
+heller `aria-label`.
+- Hamburger krever nå at du gir den en ID og setter `aria-controls` med IDen
+til menyen hamburgeren styrer. Dette er for å gjøre det enklere å bygge en
+tilgjengelig hamburgermeny.
+
+## 10.4.0 (2021-12-21)
+
+### Features
+
+- lanser expand-button ([c86d489](https://github.com/fremtind/jokul/commit/c86d4896f714271e407b85fc473eea7b8af549fb))
+- legg til bold som egen klasse ([2837d10](https://github.com/fremtind/jokul/commit/2837d109ea94f383fabaf0ad2cbb605facc6faa9))
+
 ## 10.3.17 (2021-12-20)
 
 ### Bug Fixes